fix: BGE-857 Lighthouse performance — image CLS and recharts lazy split
- Add width/height/fetchpriority attrs to avatar <img> in PlayerProfile
  to prevent layout shift (CLS) and hint browser for early LCP load
- Extract recharts chart components from PlayerStats into a separate
  PlayerStatsCharts module, lazy-imported with Suspense so recharts
  (~300KB) is never fetched for users with no game history
- Verified: ResourceHistoryChart is already lazily imported in Base.tsx
- Verified: ReplayViewer is already lazy-loaded in App.tsx
